+++ rest-service/src/test/java/org/ow2/authzforce/rest/service/test/DomainTest.java
@@ -43,7 +43,7 @@ import oasis.names.tc.xacml._3_0.core.schema.wd_17.Target;
 import org.apache.cxf.jaxrs.utils.schemas.SchemaHandler;
 import org.ow2.authzforce.core.test.utils.TestUtils;
 import org.ow2.authzforce.core.xmlns.test.TestAttributeProvider;
-import org.ow2.authzforce.pap.dao.file.FileBasedDAOUtils;
+import org.ow2.authzforce.pap.dao.flatfile.FlatFileDAOUtils;
 import org.ow2.authzforce.rest.api.jaxrs.AttributeProvidersResource;
 import org.ow2.authzforce.rest.api.jaxrs.DomainPropertiesResource;
 import org.ow2.authzforce.rest.api.jaxrs.DomainResource;
@@ -76,9 +76,50 @@ import org.w3._2005.atom.Link;
 /**
  * Tests specific to a domain
  * <p>
- * TODO: tests for max number of policies per domain, and max number of versions
- * per policy.
- *
+ * TODO:
+ * <p>
+ * - test for max number of policies per domain
+ * <p>
+ * - test for max number of versions per policy
+ * <p>
+ * - API allows the special keyword "latest" as version ID to get the latest version of a given policy (in addition to
+ * XACML version IDs like before), e.g. URL path /domains/{domainId}/pap/policies/P1/latest represents the latest
+ * version of policy "P1"
+ * <p>
+ * - Path /domains/{domainId}/pap/status giving the status of the PDP (date/time of last modification and active
+ * policies)
+ * <p>
+ * - Option for automatic removal of oldest versions if maximum allowed number of versions for a policy is reached
+ * <p>
+ * - Manual synchronization of domain cache with data directory via REST API, allows to force reloading domains' PDPs
+ * and externalIDs without restarting the webapp or server:
+ * <p>
+ * - GET /domains forces re-synchronization of all domains (test after multiple deletions and additions on disk, test
+ * also external IDs)
+ * <p>
+ * - GET /domain/{domainId}/properties forces re-synchronization of externalId with domain properties file
+ * (properties.xml) in the domain directory (test also externalId)
+ * <p>
+ * - GET /domain/{domainId}/pap/properties forces re-synchronization of PDP with configuration file (pdp.xml) in the
+ * domain directory (test change of rootPolicyRef)
+ * <p>
+ * - GET /domain/{domainId}/pap/policies forces re-synchronization of PDP with policy files in subfolder 'policies' of
+ * the domain directory (test /pap/properties to see what changed in cache)
+ * <p>
+ * - DELETE /domains/{id} removes a domain even if directory already removed
+ * <p>
+ * AUTOMATIC SYNC:
+ * <p>
+ * Test domain directory creation, deletion, update (externalId, pdp.xml, one of active policies):
+ * <p>
+ * - If a given domain ID is requested and no matching domain in cache, but a matching domain directory is found, the
+ * domain is automatically synced to cache and the synchronizing thread created;
+ * <p>
+ * - If the domain's directory found missing by the synchronizing thread, the thread deletes the domain from cache.
+ * <p>
+ * - If any change to properties.xml (domain description, externalId) detected, externalId updated in cache
+ * <p>
+ * - If any change to pdp.xml or the file of any policy used by the PDP, the PDP is reloaded.
  */
 @ContextConfiguration(locations = { "classpath:META-INF/spring/client.xml" })
 public class DomainTest extends AbstractTestNGSpringContextTests
@@ -87,8 +128,7 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 
 	private static final String ATTRIBUTE_PROVIDER_FILENAME = "attributeProvider.xml";
 
-	private static final Logger LOGGER = LoggerFactory
-			.getLogger(DomainTest.class);
+	private static final Logger LOGGER = LoggerFactory.getLogger(DomainTest.class);
 
 	private static final Random PRNG = new Random();
 
@@ -116,20 +156,17 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 	{
 		try
 		{
-			JAXB_CTX = JAXBContext.newInstance(PolicySet.class, Request.class,
-					Resources.class, TestAttributeProvider.class);
+			JAXB_CTX = JAXBContext.newInstance(PolicySet.class, Request.class, Resources.class,
+					TestAttributeProvider.class);
 		} catch (JAXBException e)
 		{
-			throw new RuntimeException(
-					"Error instantiating JAXB context for XML to Java binding",
-					e);
+			throw new RuntimeException("Error instantiating JAXB context for XML to Java binding", e);
 		}
 	}
 
 	/**
-	 * XACML policy filename used by default when no PDP configuration file
-	 * found, i.e. no file named {@value #PDP_CONF_FILENAME} exists in the test
-	 * directory
+	 * XACML policy filename used by default when no PDP configuration file found, i.e. no file named
+	 * {@value #PDP_CONF_FILENAME} exists in the test directory
 	 */
 	public final static String POLICY_FILENAME = "policy.xml";
 
@@ -156,33 +193,27 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 	private String testDomainId = null;
 
 	@BeforeClass
-	public void addDomain(ITestContext testCtx) throws JAXBException,
-			IOException
+	public void addDomain(ITestContext testCtx) throws JAXBException, IOException
 	{
 		/*
-		 * WARNING: if tests are to be multi-threaded, modify according to
-		 * Thread-safety section of CXF JAX-RS client API documentation
-		 * http://cxf
-		 * .apache.org/docs/jax-rs-client-api.html#JAX-RSClientAPI-ThreadSafety
+		 * WARNING: if tests are to be multi-threaded, modify according to Thread-safety section of CXF JAX-RS client
+		 * API documentation http://cxf .apache.org/docs/jax-rs-client-api.html#JAX-RSClientAPI-ThreadSafety
 		 */
 		final DomainsResource client = (DomainsResource) testCtx
 				.getAttribute(DomainSetTest.REST_CLIENT_TEST_CONTEXT_ATTRIBUTE_ID);
-		final Link domainLink = client.addDomain(new DomainProperties(
-				"Some description", EXTERNAL_ID));
+		final Link domainLink = client.addDomain(new DomainProperties("Some description", EXTERNAL_ID));
 		assertNotNull(domainLink, "Domain creation failure");
 
 		// The link href gives the new domain ID
 		testDomainId = domainLink.getHref();
 		LOGGER.debug("Added domain ID={}", testDomainId);
 		testDomain = client.getDomainResource(testDomainId);
-		assertNotNull(testDomain,
-				String.format("Error retrieving domain ID=%s", testDomainId));
+		assertNotNull(testDomain, String.format("Error retrieving domain ID=%s", testDomainId));
 
 		// Unmarshall test XACML PolicySet
 		if (!xacmlSamplesDirResource.exists())
 		{
-			throw new RuntimeException("Test data directory '"
-					+ xacmlSamplesDirResource + "' does not exist");
+			throw new RuntimeException("Test data directory '" + xacmlSamplesDirResource + "' does not exist");
 		}
 
 		xacmlSamplesDir = xacmlSamplesDirResource.getFile();
@@ -197,8 +228,7 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 	 */
 	public void deleteDomain()
 	{
-		assertNotNull(testDomain.deleteDomain(),
-				String.format("Error deleting domain ID=%s", testDomainId));
+		assertNotNull(testDomain.deleteDomain(), String.format("Error deleting domain ID=%s", testDomainId));
 	}
 
 	@Test
@@ -214,8 +244,7 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 	@Test
 	public void getDomainProperties()
 	{
-		final DomainPropertiesResource propsResource = testDomain
-				.getDomainPropertiesResource();
+		final DomainPropertiesResource propsResource = testDomain.getDomainPropertiesResource();
 		assertNotNull(propsResource);
 
 		final DomainProperties props = propsResource.getDomainProperties();
@@ -225,23 +254,19 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 	@Test(dependsOnMethods = { "getDomainProperties" })
 	public void updateDomainProperties()
 	{
-		final DomainPropertiesResource propsResource = testDomain
-				.getDomainPropertiesResource();
+		final DomainPropertiesResource propsResource = testDomain.getDomainPropertiesResource();
 		assertNotNull(propsResource);
 
 		byte[] randomBytes = new byte[128];
 		PRNG.nextBytes(randomBytes);
-		String description = FileBasedDAOUtils.base64UrlEncode(randomBytes);
+		String description = FlatFileDAOUtils.base64UrlEncode(randomBytes);
 		byte[] randomBytes2 = new byte[16];
 		PRNG.nextBytes(randomBytes2);
 		// externalId must be a xs:NCName, therefore cannot start with a number
-		String externalId = "external"
-				+ FileBasedDAOUtils.base64UrlEncode(randomBytes2);
-		propsResource.updateDomainProperties(new DomainProperties(description,
-				externalId));
+		String externalId = "external" + FlatFileDAOUtils.base64UrlEncode(randomBytes2);
+		propsResource.updateDomainProperties(new DomainProperties(description, externalId));
 		// verify result
-		final DomainProperties newProps = testDomain
-				.getDomainPropertiesResource().getDomainProperties();
+		final DomainProperties newProps = testDomain.getDomainPropertiesResource().getDomainProperties();
 		assertEquals(newProps.getDescription(), description);
 		assertEquals(newProps.getExternalId(), externalId);
 	}
@@ -249,16 +274,14 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 	@Test(dependsOnMethods = { "getDomain" })
 	public void getPap()
 	{
-		final ResourceContent papResource = testDomain.getPapResource()
-				.getPAP();
+		final ResourceContent papResource = testDomain.getPapResource().getPAP();
 		assertNotNull(papResource);
 	}
 
 	@Test(dependsOnMethods = { "getPap" })
 	public void getAttributeProviders()
 	{
-		final AttributeProviders attributeProviders = testDomain
-				.getPapResource().getAttributeProvidersResource()
+		final AttributeProviders attributeProviders = testDomain.getPapResource().getAttributeProvidersResource()
 				.getAttributeProviderList();
 		assertNotNull(attributeProviders);
 	}
@@ -267,27 +290,20 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 	@Test(dependsOnMethods = { "getAttributeProviders" })
 	public void updateAttributeProviders() throws JAXBException
 	{
-		final AttributeProvidersResource attributeProvidersResource = testDomain
-				.getPapResource().getAttributeProvidersResource();
-		JAXBElement<TestAttributeProvider> jaxbElt = (JAXBElement<TestAttributeProvider>) JAXB_CTX
-				.createUnmarshaller().unmarshal(
-						new File(xacmlSamplesDir,
-								"pdp/IIA002(PolicySet)/attributeProvider.xml"));
+		final AttributeProvidersResource attributeProvidersResource = testDomain.getPapResource()
+				.getAttributeProvidersResource();
+		JAXBElement<TestAttributeProvider> jaxbElt = (JAXBElement<TestAttributeProvider>) JAXB_CTX.createUnmarshaller()
+				.unmarshal(new File(xacmlSamplesDir, "pdp/IIA002(PolicySet)/attributeProvider.xml"));
 		TestAttributeProvider testAttrProvider = jaxbElt.getValue();
 		AttributeProviders updateAttrProvidersResult = attributeProvidersResource
-				.updateAttributeProviderList(new AttributeProviders(
-						Collections
-								.<AbstractAttributeProvider> singletonList(testAttrProvider)));
+				.updateAttributeProviderList(new AttributeProviders(Collections
+						.<AbstractAttributeProvider> singletonList(testAttrProvider)));
 		assertNotNull(updateAttrProvidersResult);
-		assertEquals(updateAttrProvidersResult.getAttributeProviders().size(),
-				1);
-		AbstractAttributeProvider updateAttrProvidersItem = updateAttrProvidersResult
-				.getAttributeProviders().get(0);
+		assertEquals(updateAttrProvidersResult.getAttributeProviders().size(), 1);
+		AbstractAttributeProvider updateAttrProvidersItem = updateAttrProvidersResult.getAttributeProviders().get(0);
 		if (updateAttrProvidersItem instanceof TestAttributeProvider)
 		{
-			assertEquals(
-					((TestAttributeProvider) updateAttrProvidersItem)
-							.getAttributes(),
+			assertEquals(((TestAttributeProvider) updateAttrProvidersItem).getAttributes(),
 					testAttrProvider.getAttributes());
 		} else
 		{
@@ -296,17 +312,13 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 		}
 
 		// check getAttributeProviders
-		AttributeProviders getAttrProvidersResult = attributeProvidersResource
-				.getAttributeProviderList();
+		AttributeProviders getAttrProvidersResult = attributeProvidersResource.getAttributeProviderList();
 		assertNotNull(getAttrProvidersResult);
 		assertEquals(getAttrProvidersResult.getAttributeProviders().size(), 1);
-		AbstractAttributeProvider getAttrProvidersItem = getAttrProvidersResult
-				.getAttributeProviders().get(0);
+		AbstractAttributeProvider getAttrProvidersItem = getAttrProvidersResult.getAttributeProviders().get(0);
 		if (getAttrProvidersItem instanceof TestAttributeProvider)
 		{
-			assertEquals(
-					((TestAttributeProvider) getAttrProvidersItem)
-							.getAttributes(),
+			assertEquals(((TestAttributeProvider) getAttrProvidersItem).getAttributes(),
 					testAttrProvider.getAttributes());
 		} else
 		{
@@ -319,59 +331,37 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 	@Test(dependsOnMethods = { "getPap" })
 	public void getPolicies()
 	{
-		final Resources resources = testDomain.getPapResource()
-				.getPoliciesResource().getPolicies();
+		final Resources resources = testDomain.getPapResource().getPoliciesResource().getPolicies();
 		assertNotNull(resources);
-		assertTrue(resources.getLinks().size() > 0,
-				"No resource for root policy found");
+		assertTrue(resources.getLinks().size() > 0, "No resource for root policy found");
 	}
 
 	@Test(dependsOnMethods = { "getDomainProperties", "getPolicies" })
 	public void getRootPolicy()
 	{
-		final IdReferenceType rootPolicyRef = testDomain.getPapResource()
-				.getPdpPropertiesResource().getPdpProperties()
+		final IdReferenceType rootPolicyRef = testDomain.getPapResource().getPdpPropertiesResource().getPdpProperties()
 				.getRootPolicyRef();
-		final List<Link> links = testDomain.getPapResource()
-				.getPoliciesResource()
-				.getPolicyResource(rootPolicyRef.getValue())
-				.getPolicyVersions().getLinks();
+		final List<Link> links = testDomain.getPapResource().getPoliciesResource()
+				.getPolicyResource(rootPolicyRef.getValue()).getPolicyVersions().getLinks();
 		assertTrue(links.size() > 0, "No root policy version found");
 	}
 
 	private static PolicySet createDumbPolicySet(String policyId, String version)
 	{
-		return new PolicySet(
-				null,
-				null,
-				null,
-				new Target(null),
-				null,
-				null,
-				null,
-				policyId,
-				version,
-				"urn:oasis:names:tc:xacml:3.0:policy-combining-algorithm:deny-unless-permit",
-				BigInteger.ZERO);
+		return new PolicySet(null, null, null, new Target(null), null, null, null, policyId, version,
+				"urn:oasis:names:tc:xacml:3.0:policy-combining-algorithm:deny-unless-permit", BigInteger.ZERO);
 	}
 
-	private static void matchPolicySets(PolicySet actual, PolicySet expected,
-			String testedMethodId)
+	private static void matchPolicySets(PolicySet actual, PolicySet expected, String testedMethodId)
 	{
 		assertEquals(
 				actual.getPolicySetId(),
 				expected.getPolicySetId(),
-				String.format(
-						"Actual PolicySetId (='%s') from %s() != expected PolicySetId (='%s')",
-						actual.getPolicySetId(), testedMethodId,
-						expected.getPolicySetId()));
-		assertEquals(
-				actual.getVersion(),
-				expected.getVersion(),
-				String.format(
-						"Actual PolicySet Version (='%s') from %s() != expected PolicySet Version (='%s')",
-						actual.getVersion(), testedMethodId,
-						expected.getVersion()));
+				String.format("Actual PolicySetId (='%s') from %s() != expected PolicySetId (='%s')",
+						actual.getPolicySetId(), testedMethodId, expected.getPolicySetId()));
+		assertEquals(actual.getVersion(), expected.getVersion(), String.format(
+				"Actual PolicySet Version (='%s') from %s() != expected PolicySet Version (='%s')",
+				actual.getVersion(), testedMethodId, expected.getVersion()));
 	}
 
 	private static boolean isHrefMatched(String href, List<Link> links)
@@ -388,47 +378,38 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 	}
 
 	/**
-	 * Adds a given PolicySet and returns allocated resource ID. It also checks
-	 * 2 things: 1) The response's PolicySet matches the input PolicySet (simply
-	 * checking PolicySet IDs and versions) 2) The reponse to a getPolicySet()
-	 * also matches the input PolicySet, to make sure the new PolicySet was
-	 * actually committed succesfully and no error occurred in the process. 3)
-	 * Response to getPolicyResources() with same PolicySetId and Version
-	 * matches.
+	 * Adds a given PolicySet and returns allocated resource ID. It also checks 2 things: 1) The response's PolicySet
+	 * matches the input PolicySet (simply checking PolicySet IDs and versions) 2) The reponse to a getPolicySet() also
+	 * matches the input PolicySet, to make sure the new PolicySet was actually committed succesfully and no error
+	 * occurred in the process. 3) Response to getPolicyResources() with same PolicySetId and Version matches.
 	 * <p>
 	 * Returns created policy resource path segment
 	 */
 	private String testAddAndGetPolicy(PolicySet policySet)
 	{
 		// put new policyset
-		final Link link = testDomain.getPapResource().getPoliciesResource()
-				.addPolicy(policySet);
+		final Link link = testDomain.getPapResource().getPoliciesResource().addPolicy(policySet);
 		final String policyHref = link.getHref();
 		final String[] parts = policyHref.split("/");
-		assertTrue(parts.length >= 2,
-				"Link returned by addPolicy does not match pattern 'policyId/version'");
+		assertTrue(parts.length >= 2, "Link returned by addPolicy does not match pattern 'policyId/version'");
 		final String policyResId = parts[0];
 		final String versionResId = parts[1];
 
 		// Check result was committed
 		// check added policy link is in policies list
-		PoliciesResource policiesRes = testDomain.getPapResource()
-				.getPoliciesResource();
-		assertTrue(
-				isHrefMatched(policyResId, policiesRes.getPolicies().getLinks()),
+		PoliciesResource policiesRes = testDomain.getPapResource().getPoliciesResource();
+		assertTrue(isHrefMatched(policyResId, policiesRes.getPolicies().getLinks()),
 				"Added policy resource link not found in links returned by getPoliciesResource()");
 
 		// check added policy version is in policy versions list
 		PolicyResource policyRes = policiesRes.getPolicyResource(policyResId);
 		final Resources policyVersionsResources = policyRes.getPolicyVersions();
-		assertTrue(
-				isHrefMatched(versionResId, policyVersionsResources.getLinks()),
+		assertTrue(isHrefMatched(versionResId, policyVersionsResources.getLinks()),
 				"Added policy version resource link not found in links returned by getPolicyVersions()");
 
 		// check PolicySet of added policy id/version is actually the one we
 		// added
-		final PolicySet getRespPolicySet = policyRes.getPolicyVersionResource(
-				versionResId).getPolicyVersion();
+		final PolicySet getRespPolicySet = policyRes.getPolicyVersionResource(versionResId).getPolicyVersion();
 		matchPolicySets(getRespPolicySet, policySet, "getPolicy");
 
 		return policyResId;
@@ -445,9 +426,8 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 
 		PolicySet policySet3 = createDumbPolicySet("testPolicyAdd", "1.2");
 		testAddAndGetPolicy(policySet3);
-		Resources policyVersionsResources = testDomain.getPapResource()
-				.getPoliciesResource().getPolicyResource("testPolicyAdd")
-				.getPolicyVersions();
+		Resources policyVersionsResources = testDomain.getPapResource().getPoliciesResource()
+				.getPolicyResource("testPolicyAdd").getPolicyVersions();
 		assertEquals(policyVersionsResources.getLinks().size(), 3);
 
 	}
@@ -455,8 +435,7 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 	@Test(dependsOnMethods = { "addAndGetPolicy" })
 	public void addConflictingPolicyVersion()
 	{
-		PolicySet policySet = createDumbPolicySet(
-				"testAddConflictingPolicyVersion", "1.0");
+		PolicySet policySet = createDumbPolicySet("testAddConflictingPolicyVersion", "1.0");
 		testAddAndGetPolicy(policySet);
 
 		// must be rejected
@@ -466,8 +445,7 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 			fail("Adding the same policy did not fail with HTTP 409 Conflict as expected");
 		} catch (ClientErrorException e)
 		{
-			assertEquals(e.getResponse().getStatus(),
-					Status.CONFLICT.getStatusCode());
+			assertEquals(e.getResponse().getStatus(), Status.CONFLICT.getStatusCode());
 		}
 	}
 
@@ -481,8 +459,8 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 		testAddAndGetPolicy(policySet2);
 
 		// delete policy (all versions)
-		PolicyResource policyResource = testDomain.getPapResource()
-				.getPoliciesResource().getPolicyResource("testPolicyDelete");
+		PolicyResource policyResource = testDomain.getPapResource().getPoliciesResource()
+				.getPolicyResource("testPolicyDelete");
 		Resources versionsResources = policyResource.deletePolicy();
 		assertNotNull(versionsResources);
 		assertTrue(isHrefMatched("1.2.3", versionsResources.getLinks()));
@@ -497,11 +475,8 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 			// OK
 		}
 
-		PoliciesResource policiesRes = testDomain.getPapResource()
-				.getPoliciesResource();
-		assertFalse(
-				isHrefMatched("testPolicyDelete", policiesRes.getPolicies()
-						.getLinks()),
+		PoliciesResource policiesRes = testDomain.getPapResource().getPoliciesResource();
+		assertFalse(isHrefMatched("testPolicyDelete", policiesRes.getPolicies().getLinks()),
 				"Deleted policy resource (all versions) is still in links returned by getPoliciesResource()");
 
 	}
@@ -516,25 +491,22 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 		testAddAndGetPolicy(policySet2);
 
 		// delete one of the versions
-		PolicyVersionResource policyVersionRes = testDomain.getPapResource()
-				.getPoliciesResource().getPolicyResource("testPolicyDelete")
-				.getPolicyVersionResource("1.2.3");
+		PolicyVersionResource policyVersionRes = testDomain.getPapResource().getPoliciesResource()
+				.getPolicyResource("testPolicyDelete").getPolicyVersionResource("1.2.3");
 		PolicySet deletedPolicy = policyVersionRes.deletePolicyVersion();
 		matchPolicySets(deletedPolicy, policySet1, "deletedAndTryGetPolicy");
 
 		try
 		{
 			policyVersionRes.getPolicyVersion();
-			org.testng.Assert
-					.fail("Policy version removal failed (resource still there");
+			org.testng.Assert.fail("Policy version removal failed (resource still there");
 		} catch (NotFoundException e)
 		{
 			// OK
 		}
 
-		Resources policyVersionsResources = testDomain.getPapResource()
-				.getPoliciesResource().getPolicyResource("testPolicyDelete")
-				.getPolicyVersions();
+		Resources policyVersionsResources = testDomain.getPapResource().getPoliciesResource()
+				.getPolicyResource("testPolicyDelete").getPolicyVersions();
 		assertEquals(policyVersionsResources.getLinks().size(), 1);
 		assertTrue(isHrefMatched("1.3.1", policyVersionsResources.getLinks()));
 
@@ -543,42 +515,34 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 	@Test(dependsOnMethods = { "getRootPolicy", "deletedAndTryGetUnusedPolicy" })
 	public void deleteRootPolicy()
 	{
-		final IdReferenceType rootPolicyRef = testDomain.getPapResource()
-				.getPdpPropertiesResource().getPdpProperties()
+		final IdReferenceType rootPolicyRef = testDomain.getPapResource().getPdpPropertiesResource().getPdpProperties()
 				.getRootPolicyRef();
 
 		// must be rejected
 		try
 		{
-			testDomain.getPapResource().getPoliciesResource()
-					.getPolicyResource(rootPolicyRef.getValue()).deletePolicy();
+			testDomain.getPapResource().getPoliciesResource().getPolicyResource(rootPolicyRef.getValue())
+					.deletePolicy();
 			fail("Wrongly accepted to remove root policy");
 		} catch (BadRequestException e)
 		{
 			// OK, expected
 		}
 
-		assertNotNull(
-				testDomain.getPapResource().getPoliciesResource()
-						.getPolicyResource(rootPolicyRef.getValue())
-						.getPolicyVersions(),
+		assertNotNull(testDomain.getPapResource().getPoliciesResource().getPolicyResource(rootPolicyRef.getValue())
+				.getPolicyVersions(),
 				"Root policy was actually removed although server return HTTP 400 for removal attempt");
 
 		// make sure rootPolicyRef unchanged
-		assertEquals(rootPolicyRef, testDomain.getPapResource()
-				.getPdpPropertiesResource().getPdpProperties()
-				.getRootPolicyRef(),
-				"rootPolicyRef changed although root policy removal rejected");
+		assertEquals(rootPolicyRef, testDomain.getPapResource().getPdpPropertiesResource().getPdpProperties()
+				.getRootPolicyRef(), "rootPolicyRef changed although root policy removal rejected");
 	}
 
 	private IdReferenceType setRootPolicy(String policyId, String version)
 	{
-		PdpPropertiesResource propsRes = testDomain.getPapResource()
-				.getPdpPropertiesResource();
-		IdReferenceType newRootPolicyRef = new IdReferenceType(policyId,
-				version, null, null);
-		propsRes.updateOtherPdpProperties(new PdpPropertiesUpdate(
-				newRootPolicyRef));
+		PdpPropertiesResource propsRes = testDomain.getPapResource().getPdpPropertiesResource();
+		IdReferenceType newRootPolicyRef = new IdReferenceType(policyId, version, null, null);
+		propsRes.updateOtherPdpProperties(new PdpPropertiesUpdate(newRootPolicyRef));
 		return newRootPolicyRef;
 	}
 
@@ -589,18 +553,14 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 		return setRootPolicy(policySet.getPolicySetId(), policySet.getVersion());
 	}
 
-	@Test(dependsOnMethods = { "addAndGetPolicy", "deleteRootPolicy",
-			"updateDomainPropertiesExceptRootPolicyRef" })
+	@Test(dependsOnMethods = { "addAndGetPolicy", "deleteRootPolicy", "updateDomainPropertiesExceptRootPolicyRef" })
 	public void updateRootPolicyRefToValidPolicy() throws JAXBException
 	{
-		PdpPropertiesResource propsRes = testDomain.getPapResource()
-				.getPdpPropertiesResource();
+		PdpPropertiesResource propsRes = testDomain.getPapResource().getPdpPropertiesResource();
 
 		// get current root policy ID
-		final IdReferenceType rootPolicyRef = propsRes.getPdpProperties()
-				.getRootPolicyRef();
-		PolicyResource oldRootPolicyRes = testDomain.getPapResource()
-				.getPoliciesResource()
+		final IdReferenceType rootPolicyRef = propsRes.getPdpProperties().getRootPolicyRef();
+		PolicyResource oldRootPolicyRes = testDomain.getPapResource().getPoliciesResource()
 				.getPolicyResource(rootPolicyRef.getValue());
 
 		// point rootPolicyRef to another one
@@ -610,8 +570,7 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 		IdReferenceType newRootPolicyRef = setRootPolicy(policySet);
 
 		// verify update
-		final PdpProperties newProps = testDomain.getPapResource()
-				.getPdpPropertiesResource().getPdpProperties();
+		final PdpProperties newProps = testDomain.getPapResource().getPdpPropertiesResource().getPdpProperties();
 		assertEquals(newProps.getRootPolicyRef(), newRootPolicyRef);
 
 		// delete previous root policy ref to see if it succeeds
@@ -621,8 +580,7 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 	@Test(dependsOnMethods = { "updateRootPolicyRefToValidPolicy" })
 	public void updateRootPolicyRefToMissingPolicy() throws JAXBException
 	{
-		PdpPropertiesResource propsRes = testDomain.getPapResource()
-				.getPdpPropertiesResource();
+		PdpPropertiesResource propsRes = testDomain.getPapResource().getPdpPropertiesResource();
 		PdpProperties props = propsRes.getPdpProperties();
 
 		// get current root policy ID
@@ -632,14 +590,12 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 		String oldRootPolicyId = rootPolicyRef.getValue();
 		String newRootPolicyId = oldRootPolicyId + ".new";
 
-		IdReferenceType newRootPolicyRef = new IdReferenceType(newRootPolicyId,
-				null, null, null);
+		IdReferenceType newRootPolicyRef = new IdReferenceType(newRootPolicyId, null, null, null);
 
 		// MUST fail
 		try
 		{
-			propsRes.updateOtherPdpProperties(new PdpPropertiesUpdate(
-					newRootPolicyRef));
+			propsRes.updateOtherPdpProperties(new PdpPropertiesUpdate(newRootPolicyRef));
 			fail("Setting rootPolicyRef to missing policy did not fail as expected");
 		} catch (BadRequestException e)
 		{
@@ -647,44 +603,36 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 		}
 
 		// make sure rootPolicyRef unchanged
-		assertEquals(rootPolicyRef, testDomain.getPapResource()
-				.getPdpPropertiesResource().getPdpProperties()
+		assertEquals(rootPolicyRef, testDomain.getPapResource().getPdpPropertiesResource().getPdpProperties()
 				.getRootPolicyRef());
 	}
 
 	@Test(dependsOnMethods = { "updateRootPolicyRefToValidPolicy" })
 	public void updateRootPolicyRefToValidVersion() throws JAXBException
 	{
-		PdpPropertiesResource propsRes = testDomain.getPapResource()
-				.getPdpPropertiesResource();
+		PdpPropertiesResource propsRes = testDomain.getPapResource().getPdpPropertiesResource();
 		PdpProperties props = propsRes.getPdpProperties();
 
-		PolicySet policySet = createDumbPolicySet(
-				"testUpdateRootPolicyRefToValidVersion", "4.4");
+		PolicySet policySet = createDumbPolicySet("testUpdateRootPolicyRefToValidVersion", "4.4");
 		testAddAndGetPolicy(policySet);
 
-		PolicySet policySet2 = createDumbPolicySet(
-				"testUpdateRootPolicyRefToValidVersion", "4.5");
+		PolicySet policySet2 = createDumbPolicySet("testUpdateRootPolicyRefToValidVersion", "4.5");
 		testAddAndGetPolicy(policySet2);
 
-		PolicySet policySet3 = createDumbPolicySet(
-				"testUpdateRootPolicyRefToValidVersion", "4.6");
+		PolicySet policySet3 = createDumbPolicySet("testUpdateRootPolicyRefToValidVersion", "4.6");
 		testAddAndGetPolicy(policySet3);
 
-		IdReferenceType newRootPolicyRef = new IdReferenceType(
-				"testUpdateRootPolicyRefToValidVersion", "4.5", null, null);
-		propsRes.updateOtherPdpProperties(new PdpPropertiesUpdate(
-				newRootPolicyRef));
+		IdReferenceType newRootPolicyRef = new IdReferenceType("testUpdateRootPolicyRefToValidVersion", "4.5", null,
+				null);
+		propsRes.updateOtherPdpProperties(new PdpPropertiesUpdate(newRootPolicyRef));
 
 		// verify update
-		final PdpProperties newProps = testDomain.getPapResource()
-				.getPdpPropertiesResource().getPdpProperties();
+		final PdpProperties newProps = testDomain.getPapResource().getPdpPropertiesResource().getPdpProperties();
 		assertEquals(newProps.getRootPolicyRef(), newRootPolicyRef);
 
 		// delete other policy versions (must succeed) to check that the root
 		// policyRef is not pointing to one of them by mistake
-		PolicyResource policyRes = testDomain.getPapResource()
-				.getPoliciesResource()
+		PolicyResource policyRes = testDomain.getPapResource().getPoliciesResource()
 				.getPolicyResource("testUpdateRootPolicyRefToValidVersion");
 		policyRes.getPolicyVersionResource("4.4").deletePolicyVersion();
 		policyRes.getPolicyVersionResource("4.6").deletePolicyVersion();
@@ -693,8 +641,7 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 	@Test(dependsOnMethods = { "updateRootPolicyRefToValidVersion" })
 	public void updateRootPolicyRefToMissingVersion() throws JAXBException
 	{
-		PdpPropertiesResource propsRes = testDomain.getPapResource()
-				.getPdpPropertiesResource();
+		PdpPropertiesResource propsRes = testDomain.getPapResource().getPdpPropertiesResource();
 		PdpProperties props = propsRes.getPdpProperties();
 
 		// get current root policy ID
@@ -702,13 +649,11 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 
 		// point rootPolicyRef to same policy id but different version
 		String rootPolicyId = rootPolicyRef.getValue();
-		IdReferenceType newRootPolicyRef = new IdReferenceType(rootPolicyId,
-				"0.0.0.1", null, null);
+		IdReferenceType newRootPolicyRef = new IdReferenceType(rootPolicyId, "0.0.0.1", null, null);
 		// MUST FAIL
 		try
 		{
-			propsRes.updateOtherPdpProperties(new PdpPropertiesUpdate(
-					newRootPolicyRef));
+			propsRes.updateOtherPdpProperties(new PdpPropertiesUpdate(newRootPolicyRef));
 			fail("Setting rootPolicyRef to missing policy version did not fail as expected.");
 		} catch (BadRequestException e)
 		{
@@ -716,21 +661,18 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 		}
 
 		// make sure rootPolicyRef unchanged
-		assertEquals(rootPolicyRef, testDomain.getPapResource()
-				.getPdpPropertiesResource().getPdpProperties()
+		assertEquals(rootPolicyRef, testDomain.getPapResource().getPdpPropertiesResource().getPdpProperties()
 				.getRootPolicyRef());
 	}
 
 	private void resetPolicies() throws JAXBException
 	{
-		final PolicySet rootPolicySet = (PolicySet) unmarshaller
-				.unmarshal(new File(xacmlSamplesDir, DEFAULT_POLICYSET_FILENAME));
-		final PoliciesResource policiesRes = testDomain.getPapResource()
-				.getPoliciesResource();
+		final PolicySet rootPolicySet = (PolicySet) unmarshaller.unmarshal(new File(xacmlSamplesDir,
+				DEFAULT_POLICYSET_FILENAME));
+		final PoliciesResource policiesRes = testDomain.getPapResource().getPoliciesResource();
 
 		// If already exists, remove to start from scratch
-		PolicyResource rootPolicyRes = policiesRes
-				.getPolicyResource(rootPolicySet.getPolicySetId());
+		PolicyResource rootPolicyRes = policiesRes.getPolicyResource(rootPolicySet.getPolicySetId());
 		try
 		{
 			rootPolicyRes.deletePolicy();
@@ -764,31 +706,27 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 	public void setRootPolicyWithGoodRefs() throws JAXBException
 	{
 		/*
-		 * Before putting the empty refPolicySets (to make a policySets with
-		 * references wrong), first put good root policySet without policy
-		 * references (in case the current root PolicySet has references,
-		 * uploading the empty refPolicySets before will be rejected because it
-		 * makes the policySet with references invalid)
+		 * Before putting the empty refPolicySets (to make a policySets with references wrong), first put good root
+		 * policySet without policy references (in case the current root PolicySet has references, uploading the empty
+		 * refPolicySets before will be rejected because it makes the policySet with references invalid)
 		 */
 		resetPolicies();
 
-		final PolicySet refPolicySet = (PolicySet) unmarshaller
-				.unmarshal(new File(xacmlSamplesDir,
-						"pdp/PolicyReference.Valid/refPolicies/pps-employee.xml"));
+		final PolicySet refPolicySet = (PolicySet) unmarshaller.unmarshal(new File(xacmlSamplesDir,
+				"pdp/PolicyReference.Valid/refPolicies/pps-employee.xml"));
 		String refPolicyResId = testAddAndGetPolicy(refPolicySet);
 
 		// Set root policy referencing ref policy above
-		final PolicySet policySetWithRef = (PolicySet) unmarshaller
-				.unmarshal(new File(xacmlSamplesDir,
-						"pdp/PolicyReference.Valid/policy.xml"));
+		final PolicySet policySetWithRef = (PolicySet) unmarshaller.unmarshal(new File(xacmlSamplesDir,
+				"pdp/PolicyReference.Valid/policy.xml"));
 		// Add the policy and point the rootPolicyRef to new policy with refs to
 		// instantiate it as root policy (validate, etc.)
 		setRootPolicy(policySetWithRef);
 
 		// Delete referenced policy -> must fail (because root policy is still
 		// referencing one of them)
-		final PolicyResource refPolicyResource = testDomain.getPapResource()
-				.getPoliciesResource().getPolicyResource(refPolicyResId);
+		final PolicyResource refPolicyResource = testDomain.getPapResource().getPoliciesResource()
+				.getPolicyResource(refPolicyResId);
 		try
 		{
 			refPolicyResource.deletePolicy();
@@ -799,11 +737,9 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 		}
 
 		/*
-		 * Check that referenced policy set was not actually deleted, is still
-		 * there
+		 * Check that referenced policy set was not actually deleted, is still there
 		 */
-		final PolicySet getPolicyVersion = refPolicyResource
-				.getPolicyVersionResource(refPolicySet.getVersion())
+		final PolicySet getPolicyVersion = refPolicyResource.getPolicyVersionResource(refPolicySet.getVersion())
 				.getPolicyVersion();
 		assertNotNull(getPolicyVersion);
 	}
@@ -812,24 +748,20 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 	public void setRootPolicyWithRefToMissingPolicy() throws JAXBException
 	{
 		/*
-		 * Before putting the empty refPolicySets (to make a policySets with
-		 * references wrong), first put good root policySet without policy
-		 * references (in case the current root PolicySet has references,
-		 * uploading the empty refPolicySets before will be rejected because it
-		 * makes the policySet with references invalid)
+		 * Before putting the empty refPolicySets (to make a policySets with references wrong), first put good root
+		 * policySet without policy references (in case the current root PolicySet has references, uploading the empty
+		 * refPolicySets before will be rejected because it makes the policySet with references invalid)
 		 */
 		resetPolicies();
 		// Get current rootPolicy
-		final IdReferenceType rootPolicyRef = testDomain.getPapResource()
-				.getPdpPropertiesResource().getPdpProperties()
+		final IdReferenceType rootPolicyRef = testDomain.getPapResource().getPdpPropertiesResource().getPdpProperties()
 				.getRootPolicyRef();
 
 		// Then attempt to put bad root policy set (referenced policysets in
 		// Policy(Set)IdReferences do
 		// not exist since refPolicySets empty)
-		final PolicySet policySetWithRefs = (PolicySet) unmarshaller
-				.unmarshal(new File(xacmlSamplesDir,
-						"PolicyReference.Undef/policy.xml"));
+		final PolicySet policySetWithRefs = (PolicySet) unmarshaller.unmarshal(new File(xacmlSamplesDir,
+				"PolicyReference.Undef/policy.xml"));
 		try
 		{
 			setRootPolicy(policySetWithRefs);
@@ -840,44 +772,36 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 		}
 
 		// make sure the rootPolicyRef is unchanged
-		assertEquals(rootPolicyRef, testDomain.getPapResource()
-				.getPdpPropertiesResource().getPdpProperties()
-				.getRootPolicyRef(),
-				"rootPolicyRef changed although root policy update rejected");
+		assertEquals(rootPolicyRef, testDomain.getPapResource().getPdpPropertiesResource().getPdpProperties()
+				.getRootPolicyRef(), "rootPolicyRef changed although root policy update rejected");
 	}
 
 	@Test(dependsOnMethods = { "setRootPolicyWithGoodRefs" })
 	public void setRootPolicyWithCircularRef() throws JAXBException
 	{
 		/*
-		 * Before putting the empty refPolicySets (to make a policySets with
-		 * references wrong), first put good root policySet without policy
-		 * references (in case the current root PolicySet has references,
-		 * uploading the empty refPolicySets before will be rejected because it
-		 * makes the policySet with references invalid)
+		 * Before putting the empty refPolicySets (to make a policySets with references wrong), first put good root
+		 * policySet without policy references (in case the current root PolicySet has references, uploading the empty
+		 * refPolicySets before will be rejected because it makes the policySet with references invalid)
 		 */
 		resetPolicies();
 		// Get current rootPolicy
-		final IdReferenceType rootPolicyRef = testDomain.getPapResource()
-				.getPdpPropertiesResource().getPdpProperties()
+		final IdReferenceType rootPolicyRef = testDomain.getPapResource().getPdpPropertiesResource().getPdpProperties()
 				.getRootPolicyRef();
 
 		// add refPolicies
-		final PolicySet refPolicySet = (PolicySet) unmarshaller
-				.unmarshal(new File(xacmlSamplesDir,
-						"PolicyReference.Circular/refPolicies/invalid-pps-employee.xml"));
+		final PolicySet refPolicySet = (PolicySet) unmarshaller.unmarshal(new File(xacmlSamplesDir,
+				"PolicyReference.Circular/refPolicies/invalid-pps-employee.xml"));
 		testAddAndGetPolicy(refPolicySet);
-		final PolicySet refPolicySet2 = (PolicySet) unmarshaller
-				.unmarshal(new File(xacmlSamplesDir,
-						"PolicyReference.Circular/refPolicies/pps-manager.xml"));
+		final PolicySet refPolicySet2 = (PolicySet) unmarshaller.unmarshal(new File(xacmlSamplesDir,
+				"PolicyReference.Circular/refPolicies/pps-manager.xml"));
 		testAddAndGetPolicy(refPolicySet2);
 
 		// Then attempt to put bad root policy set (referenced policysets in
 		// Policy(Set)IdReferences do
 		// not exist since refPolicySets empty)
-		final PolicySet policySetWithRefs = (PolicySet) unmarshaller
-				.unmarshal(new File(xacmlSamplesDir,
-						"PolicyReference.Circular/policy.xml"));
+		final PolicySet policySetWithRefs = (PolicySet) unmarshaller.unmarshal(new File(xacmlSamplesDir,
+				"PolicyReference.Circular/policy.xml"));
 		try
 		{
 			setRootPolicy(policySetWithRefs);
@@ -888,10 +812,8 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 		}
 
 		// make sure the rootPolicyRef is unchanged
-		assertEquals(rootPolicyRef, testDomain.getPapResource()
-				.getPdpPropertiesResource().getPdpProperties()
-				.getRootPolicyRef(),
-				"rootPolicyRef changed although root policy update rejected");
+		assertEquals(rootPolicyRef, testDomain.getPapResource().getPdpPropertiesResource().getPdpProperties()
+				.getRootPolicyRef(), "rootPolicyRef changed although root policy update rejected");
 	}
 
 	@Test(dependsOnMethods = { "addAndGetPolicy" }, expectedExceptions = NotFoundException.class)
@@ -900,23 +822,20 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 		// Then attempt to put bad root policy set (referenced policysets in
 		// Policy(Set)IdReferences do
 		// not exist since refPolicySets empty)
-		final PolicySet badPolicySet = (PolicySet) unmarshaller
-				.unmarshal(new File(xacmlSamplesDir,
-						"policySetWithTooManyChildElements.xml"));
+		final PolicySet badPolicySet = (PolicySet) unmarshaller.unmarshal(new File(xacmlSamplesDir,
+				"policySetWithTooManyChildElements.xml"));
 		try
 		{
 			testAddAndGetPolicy(badPolicySet);
 			fail("Invalid PolicySet (too many child elements) accepted");
 		} catch (ClientErrorException e)
 		{
-			assertEquals(e.getResponse().getStatus(),
-					Status.REQUEST_ENTITY_TOO_LARGE.getStatusCode());
+			assertEquals(e.getResponse().getStatus(), Status.REQUEST_ENTITY_TOO_LARGE.getStatusCode());
 		}
 
 		// make sure the policy is not there
 		// MUST throw NotFoundException
-		testDomain.getPapResource().getPoliciesResource()
-				.getPolicyResource(badPolicySet.getPolicySetId())
+		testDomain.getPapResource().getPoliciesResource().getPolicyResource(badPolicySet.getPolicySetId())
 				.getPolicyVersions();
 	}
 
@@ -926,8 +845,8 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 		// Then attempt to put bad root policy set (referenced policysets in
 		// Policy(Set)IdReferences do
 		// not exist since refPolicySets empty)
-		final PolicySet badPolicySet = (PolicySet) unmarshaller
-				.unmarshal(new File(xacmlSamplesDir, "policySetTooDeep.xml"));
+		final PolicySet badPolicySet = (PolicySet) unmarshaller.unmarshal(new File(xacmlSamplesDir,
+				"policySetTooDeep.xml"));
 		try
 		{
 			testAddAndGetPolicy(badPolicySet);
@@ -939,14 +858,12 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 
 		// make sure the policy is not there
 		// MUST throw NotFoundException
-		testDomain.getPapResource().getPoliciesResource()
-				.getPolicyResource(badPolicySet.getPolicySetId())
+		testDomain.getPapResource().getPoliciesResource().getPolicyResource(badPolicySet.getPolicySetId())
 				.getPolicyVersions();
 	}
 
 	/**
-	 * Create PDP evaluation test data. Various PolicySets/Requests/Responses
-	 * from conformance tests.
+	 * Create PDP evaluation test data. Various PolicySets/Requests/Responses from conformance tests.
 	 * 
 	 * 
 	 * @return iterator over test data
@@ -954,13 +871,12 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 	 * @throws IOException
 	 */
 	@DataProvider(name = "pdpTestFiles")
-	public Iterator<Object[]> createData() throws URISyntaxException,
-			IOException
+	public Iterator<Object[]> createData() throws URISyntaxException, IOException
 	{
 		final Collection<Object[]> testParams = new ArrayList<>();
 		/*
-		 * Each sub-directory of the root directory is data for a specific test.
-		 * So we configure a test for each directory
+		 * Each sub-directory of the root directory is data for a specific test. So we configure a test for each
+		 * directory
 		 */
 		final File testRootDir = new File(xacmlSamplesDir, "pdp");
 		for (final File subDir : testRootDir.listFiles(DIRECTORY_FILTER))
@@ -973,17 +889,13 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 		return testParams.iterator();
 	}
 
-	private static void assertNormalizedEquals(Response expectedResponse,
-			Response actualResponseFromPDP) throws JAXBException
+	private static void assertNormalizedEquals(Response expectedResponse, Response actualResponseFromPDP)
+			throws JAXBException
 	{
 		// normalize responses for comparison
-		final Response normalizedExpectedResponse = TestUtils
-				.normalizeForComparison(expectedResponse);
-		final Response normalizedActualResponse = TestUtils
-				.normalizeForComparison(actualResponseFromPDP);
-		assertEquals(
-				normalizedExpectedResponse,
-				normalizedActualResponse,
+		final Response normalizedExpectedResponse = TestUtils.normalizeForComparison(expectedResponse);
+		final Response normalizedActualResponse = TestUtils.normalizeForComparison(actualResponseFromPDP);
+		assertEquals(normalizedExpectedResponse, normalizedActualResponse,
 				"Actual and expected responses don't match (Status elements removed/ignored for comparison)");
 	}
 
@@ -996,13 +908,11 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 				.getPapResource()
 				.getAttributeProvidersResource()
 				.updateAttributeProviderList(
-						new AttributeProviders(Collections
-								.<AbstractAttributeProvider> emptyList()));
+						new AttributeProviders(Collections.<AbstractAttributeProvider> emptyList()));
 
 		LOGGER.debug("Starting PDP test of directory '{}'", testDirectory);
 
-		final File attributeProviderFile = new File(testDirectory,
-				ATTRIBUTE_PROVIDER_FILENAME);
+		final File attributeProviderFile = new File(testDirectory, ATTRIBUTE_PROVIDER_FILENAME);
 		if (attributeProviderFile.exists())
 		{
 			JAXBElement<AbstractAttributeProvider> jaxbElt = (JAXBElement<AbstractAttributeProvider>) JAXB_CTX
@@ -1011,40 +921,33 @@ public class DomainTest extends AbstractTestNGSpringContextTests
 					.getPapResource()
 					.getAttributeProvidersResource()
 					.updateAttributeProviderList(
-							new AttributeProviders(
-									Collections
-											.<AbstractAttributeProvider> singletonList(jaxbElt
-													.getValue())));
+							new AttributeProviders(Collections.<AbstractAttributeProvider> singletonList(jaxbElt
+									.getValue())));
 		}
 
-		final File refPoliciesDir = new File(testDirectory,
-				REF_POLICIES_DIRECTORY_NAME);
+		final File refPoliciesDir = new File(testDirectory, REF_POLICIES_DIRECTORY_NAME);
 		if (refPoliciesDir.exists() && refPoliciesDir.isDirectory())
 		{
 			for (final File policyFile : refPoliciesDir.listFiles())
 			{
 				if (policyFile.isFile())
 				{
-					final PolicySet policy = (PolicySet) unmarshaller
-							.unmarshal(policyFile);
+					final PolicySet policy = (PolicySet) unmarshaller.unmarshal(policyFile);
 					testAddAndGetPolicy(policy);
 					LOGGER.debug("Added policy from file: " + policyFile);
 				}
 			}
 		}
 
-		final PolicySet policy = (PolicySet) unmarshaller.unmarshal(new File(
-				testDirectory, POLICY_FILENAME));
+		final PolicySet policy = (PolicySet) unmarshaller.unmarshal(new File(testDirectory, POLICY_FILENAME));
 		// Add the policy and point the rootPolicyRef to new policy with refs to
 		// instantiate it as root policy (validate, etc.)
 		setRootPolicy(policy);
 
-		final Request xacmlReq = (Request) unmarshaller.unmarshal(new File(
-				testDirectory, REQUEST_FILENAME));
-		final Response expectedResponse = (Response) unmarshaller
-				.unmarshal(new File(testDirectory, EXPECTED_RESPONSE_FILENAME));
-		final Response actualResponse = testDomain.getPdpResource()
-				.requestPolicyDecision(xacmlReq);
+		final Request xacmlReq = (Request) unmarshaller.unmarshal(new File(testDirectory, REQUEST_FILENAME));
+		final Response expectedResponse = (Response) unmarshaller.unmarshal(new File(testDirectory,
+				EXPECTED_RESPONSE_FILENAME));
+		final Response actualResponse = testDomain.getPdpResource().requestPolicyDecision(xacmlReq);
 		assertNormalizedEquals(expectedResponse, actualResponse);
 	}